refactor: seperate apps & layout
