+++ apps/web/src/app/[lang]/(core)/management/[domain]/[data]/page.tsx
