+++ apps/web/src/app/[lang]/(core)/management/[domain]/[data]/table-actions/claims.tsx
