+++ apps/web/src/app/[lang]/(core)/management/[domain]/[data]/table-actions/permissions.tsx
