+++ apps/web/src/app/[lang]/(core)/management/[domain]/[data]/table-actions/sessions.tsx
