+++ apps/web/src/app/[lang]/(core)/settings/tenant/[group]/_components/description.tsx
