+++ apps/web/src/app/[lang]/(core)/settings/tenant/[group]/mock-settings-response.ts
