+++ apps/web/src/app/[lang]/(core)/settings/tenant/[group]/page.tsx
