+++ apps/web/src/app/[lang]/(core)/settings/tenant/page.tsx
